Moving InputPositionControlling from extension to class itself to make the methods overridable
